Fix profanity filter + new username generator
The profanity filter was fixed up to account for cases like [C]rap and C_rap. The filter was extended slightly, and the `RegExpMatcher` ctor was fixed (transformers can't use the same `...` syntax like `englishDataset` and `englishRecommendedTransformers`

Additionally, the scattered set of name generators used for players was replaced with one single name generator that comes up with some snazzy stuff like "Space Cow", "Secret Gnome Agenda", and "Comically Large Snail".

Copyright: No AI was used in the making of this patch.